azure_access_webapp/azure_access_webapp/requirements.txt
flask
python-dotenv
azure-identity
azure-ai-projects
